costs: move s3_costs package to s3/costs
This commit moves the s3_costs package to s3/costs in order to respect the go naming convention and allow for an easier extension of the s3 package
