+++ s3/costs/es_request_constructor.go
@@ -12,8 +12,8 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-// Package s3_costs gets billing information from an ElasticSearch.
-package s3_costs
+// Package s3/costs gets billing information from an ElasticSearch.
+package costs
 
 import (
 	"fmt"
