+++ s3/costs/es_request_constructor_test.go
@@ -12,7 +12,7 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-package s3_costs
+package costs
 
 import (
 	"encoding/json"
