+++ s3/costs/s3_costs_route.go
@@ -12,7 +12,7 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-package s3_costs
+package costs
 
 import (
 	"context"
