+++ server/server.go
@@ -26,7 +26,7 @@ import (
 	"github.com/trackit/trackit2/config"
 	_ "github.com/trackit/trackit2/costs"
 	"github.com/trackit/trackit2/routes"
-	_ "github.com/trackit/trackit2/s3_costs"
+	_ "github.com/trackit/trackit2/s3/costs"
 	_ "github.com/trackit/trackit2/users"
 )
